Add chrome 2 configurations.
