+++ config/webpack-ts-overrides.js
@@ -1,12 +1,33 @@
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const { resolve } = require('path');
-
+const path = require('path');
 const isBuild = process.env.NODE_ENV === 'production';
+
+const appEntry = path.resolve(
+  __dirname,
+  `../src/${isBuild ? 'entry.js' : 'entry-dev.js'}`,
+);
 
 module.exports = {
   devtool: 'source-map',
   module: {
     rules: [
+      /**
+       * Chrome render loader blocks the ReactDOM.render call if the application is loaded within the platfrom.
+       * This is because the platform will eventually only use one main HTML file generated by chrome.
+       * Thus it will not use the main output file of webpack, but the remote container entry file which does not include the rende call.
+       * Usually it is included in the shared webpack config but the override config, overrides it.
+       */
+      {
+        test: new RegExp(appEntry),
+        loader: path.resolve(
+          __dirname,
+          '../node_modules/@redhat-cloud-services/frontend-components-config-utilities/chrome-render-loader',
+        ),
+        options: {
+          appName: 'automation-hub',
+        },
+      },
       {
         test: /src\/.*\.(js|jsx|ts|tsx)$/,
         use: [{ loader: 'source-map-loader' }, { loader: 'babel-loader' }],
