+++ config/webpack.base.config.js
@@ -4,6 +4,7 @@ const config = require('@redhat-cloud-services/frontend-components-config');
 const TSOverrides = require('./webpack-ts-overrides');
 const webpack = require('webpack');
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
+const isBuild = process.env.NODE_ENV === 'production';
 
 // NOTE: This file is not meant to be consumed directly by weback. Instead it
 // should be imported, initialized with the following settings and exported like
@@ -68,7 +69,6 @@ module.exports = (inputConfigs) => {
     htmlPlugin: htmlPluginConfig,
     debug: customConfigs.UI_DEBUG,
     https: customConfigs.UI_USE_HTTPS,
-
     // defines port for dev server
     port: customConfigs.UI_PORT,
   });
@@ -98,6 +98,25 @@ module.exports = (inputConfigs) => {
 
   plugins.push(new webpack.DefinePlugin(globals));
   plugins.push(new ForkTsCheckerWebpackPlugin());
+
+  /**
+   * Generates remote containers for chrome 2
+   */
+  plugins.push(
+    require('@redhat-cloud-services/frontend-components-config/federated-modules')(
+      {
+        root: resolve(__dirname, '../'),
+        exposes: {
+          './RootApp': resolve(
+            __dirname,
+            isBuild ? '../src/AppEntry' : '../src/DevEntry',
+          ),
+        },
+      },
+    ),
+  );
+
+  console.log({ rules: newWebpackConfig.module.rules });
 
   return {
     ...newWebpackConfig,
