+++ src/boostrap-dev.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import AnsibleHub from './DevEntry';
+
+ReactDOM.render(<AnsibleHub />, document.getElementById('root'));
