+++ src/boostrap.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import AnsibleHub from './AppEntry';
+
+ReactDOM.render(<AnsibleHub />, document.getElementById('root'));
